--- src/CheXbert/src/labeled_reports.csv
@@ -1,7 +1,0 @@
-Report Impression,Enlarged Cardiomediastinum,Cardiomegaly,Lung Opacity,Lung Lesion,Edema,Consolidation,Pneumonia,Atelectasis,Pneumothorax,Pleural Effusion,Pleural Other,Fracture,Support Devices,No Finding
-Heart size normal and lungs are clear. No edema or pneumonia. No effusion,,0.0,,,0.0,,0.0,,,0.0,,,,1.0
-"1. Left pleural effusion with adjacent atelectasis. Right effusion is also present.
-
-2. Cardiomegaly without overt edema.",,1.0,,,0.0,,,1.0,,1.0,,,,
-"Minimal patchy airspace disease within the lingula, may reflect atelectasis or consolidation.",,,1.0,,,-1.0,,-1.0,,,,,,
-1. Stable mild cardiomegaly. 2. Hyperexpanded but clear lungs.,,1.0,,,,,,,,,,,,
