--- src/CheXbert/src/sample_reports.csv
@@ -1,7 +1,0 @@
-Report Impression
-Heart size normal and lungs are clear. No edema or pneumonia. No effusion
-"1. Left pleural effusion with adjacent atelectasis. Right effusion is also present.
-
-2. Cardiomegaly without overt edema."
-"Minimal patchy airspace disease within the lingula, may reflect atelectasis or consolidation."
-1. Stable mild cardiomegaly. 2. Hyperexpanded but clear lungs.
